hw/requirements.txt
streamlit
streamlit-drawable-canvas
pillow
numpy
rapidfuzz
easyocr
torch
torchvision